scaffolder-backend: add Publishers.fromConfig
